Update Readme The MyIterator success look like example was misleading.
It is impossible to use directly (0..5).my_iter() and not really a good
exercice. Range already impl Iterator. I won't be able to impl
MyIterator without adding a new type. I could but I already did it with
Vec which is good enough. It was a good exercice though to have
understand that.